# lib/api-zod/src/generated/types/adminStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsResponseCategoryBreakdownItem } from "./adminStatsResponseCategoryBreakdownItem";
import type { AdminStatsResponseTopApisItem } from "./adminStatsResponseTopApisItem";
import type { HistoryEntry } from "./historyEntry";

export interface AdminStatsResponse {
  totalLookups: number;
  successfulLookups: number;
  failedLookups: number;
  activeApis: number;
  totalApis: number;
  cachedResults: number;
  categoryBreakdown: AdminStatsResponseCategoryBreakdownItem[];
  topApis: AdminStatsResponseTopApisItem[];
  recentActivity: HistoryEntry[];
}