# lib/api-zod/src/generated/types/lookupResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { LookupResultData } from "./lookupResultData";

export interface LookupResult {
  data: LookupResultData;
  cached: boolean;
  apiName: string;
  success: boolean;
  developer: string;
  error?: string | null;
}